Sage: switch various longs to ints

+++ sage/src/main/java/com/hartwig/hmftools/sage/candidate/Candidate.java
@@ -75,8 +75,6 @@ public class Candidate
         return mVariant.chromosome();
     }
 
-    public long position()
-    {
-        return mVariant.position();
-    }
+    public int position() { return (int)mVariant.position(); }
+
 }
